added main file for the game
